+++ home-manager-modules/agenix/default.nix
@@ -1,0 +1,4 @@
+{ self, ... }:
+{
+  imports = [ self.inputs.agenix.homeManagerModules.default ];
+}