+++ home-manager-modules/ssh/default.nix
@@ -48,6 +48,7 @@ let
           AddKeysToAgent = "yes";
           ConnectTimeout = "3";
         };
+
         forwardAgent = true;
         identitiesOnly = true;
         hostname = x;