+++ hosts/alakazam/disk-config.nix
@@ -1,0 +1,57 @@
+{ self, ... }:
+{
+  imports = [ self.inputs.disko.nixosModules.default ];
+
+  fileSystems."/persistent".neededForBoot = true;
+
+  disko = {
+    enableConfig = true;
+
+    devices = {
+      nodev."/" = {
+        fsType = "tmpfs";
+        mountOptions = [
+          "size=1G"
+          "defaults"
+          "mode=755"
+        ];
+      };
+
+      disk = {
+        main = {
+          type = "disk";
+          device = "/dev/nvme0n1";
+          content = {
+            type = "gpt";
+            partitions = {
+              ESP = {
+                size = "500M";
+                type = "EF00";
+                content = {
+                  type = "filesystem";
+                  format = "vfat";
+                  mountpoint = "/boot";
+                  mountOptions = [ "umask=0077" ];
+                };
+              };
+              luks = {
+                size = "100%";
+                content = {
+                  type = "luks";
+                  name = "crypted";
+                  settings.allowDiscards = true;
+                  content = {
+                    type = "filesystem";
+                    format = "ext4";
+                    mountpoint = "/persistent";
+                    mountOptions = [ "noatime" ];
+                  };
+                };
+              };
+            };
+          };
+        };
+      };
+    };
+  };
+}