+++ hosts/gastly/default.nix
@@ -48,34 +48,13 @@ in
   ];
 
   age = {
-    secrets = {
-      git-signing-key = rec {
-        file = ../../secrets/ssh/git-signing-key.age;
-        owner = builtins.head (builtins.attrNames jay.users.users);
-        path = "/home/${owner}/.ssh/git-signing-key";
-      };
-
-      "git-signing-key.pub" = rec {
-        file = ../../secrets/ssh/git-signing-key.pub.age;
-        owner = builtins.head (builtins.attrNames jay.users.users);
-        path = "/home/${owner}/.ssh/git-signing-key.pub";
-      };
-
-      terraform-api-key = rec {
-        file = ../../secrets/terraform/terraform-api-key.age;
-        owner = builtins.head (builtins.attrNames jay.users.users);
-        mode = "400";
-        path = "/home/${owner}/.terraform.d/credentials.tfrc.json";
-      };
-
-      wireless-env = {
-        file = ../../secrets/wireless/wireless-home.env.age;
-        mode = "0400";
-      };
+    secrets.wireless-env = {
+      file = ../../secrets/wireless/wireless-home.env.age;
+      mode = "0400";
     };
+
     identityPaths = [
       "/agenix/id-ed25519-ssh-primary"
-      "/agenix/id-ed25519-terraform-primary"
       "/agenix/id-ed25519-wireless-primary"
     ];
   };