+++ hosts/lavender-tower/cloyster-linux/default.nix
@@ -21,23 +21,6 @@ in
 {
   inherit (user-configs) users home-manager;
 
-  age = {
-    secrets = {
-      "git-signing-key" = rec {
-        file = ../../secrets/ssh/git-signing-key.age;
-        owner = builtins.head (builtins.attrNames jay.users.users);
-        path = "/home/${owner}/.ssh/git-signing-key";
-      };
-
-      "git-signing-key.pub" = rec {
-        file = ../../secrets/ssh/git-signing-key.pub.age;
-        owner = builtins.head (builtins.attrNames jay.users.users);
-        path = "/home/${owner}/.ssh/git-signing-key.pub";
-      };
-    };
-    identityPaths = [ "/agenix/id-ed25519-ssh-primary" ];
-  };
-
   boot = {
     extraModulePackages = [ ];
     initrd = {